Major refactor of ModuleFuelTanks.
The core of ModuleFuelTanks is now in an appropriately named file. Also,
much of the loading code has been cleaned up. As a nice effect, loading my
type-switched tank now seems to work properly.

The GUI code is split into its own class. It is a little flaky, but the
code is much cleaner.

Engine handling is currently missing, but will be back soon in its own
class.

Plenty of reformatting of the code, too, just so I can read it (thus why
git can't recognize the content).