Make some progress on getting load/save working.
Still not there yet, but it turns out a lot of the trouble is caused by
trying to do too much in OnLoad(). OnLoad() is meant for pulling data out
of the config node and into the appropriate data structures, not for
messing around in KSP. That is what OnStart() is for.