Do not remove resources that MFT does not manage.
A part's resources that MFT manages are recorded in a database and when the
tank type is updated, only those resources that are managed by MFT and are
not in the new tank type will be removed.

This fixes the problem of MFT removing electric charge and monoprop when
TAC-LS is installed.