Add gimbal support to MEC
Both CONFIG and TL-based gimbal changing.
For CONIFG, specify gimbalRange in the CONFIG and it will be applied on
SetConfig.
For TLs, specify gimbalRange in a TL node, or TLGIMBALn values in the
global techlevels. Then, either specify which gimbal (based on
gimbalTransformName) to apply to (along with techLevel=foo,
gimbalTransform = bar) or, if you just want it applied to the first,
specify useGimbalAnyway = true; otherwise TL gimbal changes won't be
applied.