Fix upgrade manager unnecessary errors
Current code makes it so that if an upgrade containing a config isn't
found, when the engine GUI is open, it spams debug messages trying to
find the upgrade.  This makes it so that it just assumes that the
upgrade is unlocked.  Not every config is going to have an upgrade node.