Fix tank copying when cloning via symmetry.
From Flow's post:
Go to VAB, create a new vehicle with a Mk1-2 command pod. attach a FLT-200
fuel tank to the side. Now right click the fuel tank, change fuel to
monopropellant with MFT. Select a symmetry mode >1-fold and attach the
tanks to the side of the command pod, right click one fuel tank. The right
click menu should disappear and some weird kind  of right-click menu should
appear in the upper left with lots of monopropellant bars. Right-clicking
on any part does not have any effect anymore, VAB has to be left and
entered again to enable right-clicking again.
---
Copying tankList in the new OnCopy (and fixing FuelTank's CreateCopy) takes
care of things nicely. The problem with CreateCopy was the MeberwiseClone
was copying the properties too and thus mangling the tank list.