Show tooltip for current engine config
The tooltip shown when hovering over a button in the engine manager GUI
shows info not in the main engine manager window (e.g. TestFlight rated
burn times, engine mass, etc.). Prior to this commit, the only ways to
compare these values between different engine configs are to use the
popup window that appears when hovering over the part in the selection
window in the editor (where TF burn times are shown separately from the
rest of the engine stats), or to click on the button for a different
engine config and hover over the button for the original config.
Neither way is particularly user-friendly, and the latter one may be
problematic if the player hadn't purchased the corresponding config yet.

Showing the tooltip for the current part makes comparisons slightly
easier by not requiring the user to select a different engine config to
compare certain values. It's still not perfect, since the tooltip and
main engine config window don't show the same info.