Stop downgrading locked engine configs on vessel load
Idea from schnobs on #RO.

Automatically downgrading from locked engine configs on vessel load
isn't particularly player-friendly, since there is no indication that
such a change has taken place. This could lead to the player unwittingly
launching with the wrong engine and/or fuel mixture.

Fixing this would entail either popping up a message box or removing the
downgrading behavior. Removing the downgrading behavior is easier/less
work, and arguably is more in line with stock behavior, which is to load
vessels with locked parts and just prevent the player from launching
them.