Missing Version file added
I noticed that Ckan was showing the data incorrectly for the newest revision that was compiled for KSP 1.4.1 and works (Very well) in even KSP 1.4.9. This caused an annoying error message on startup as well as having to force overrides in Ckan to load the mod. So, I created a "version" file based on the change notes listed in the thread as well as updated the download link for GitHub and the forum posing "Home" URL. Everything shows up in Ckan correctly now. Yay!