Fixed spammy duplicate engine propellant button bug
* Fixed spammy duplicate engine propellant button bug. (for real this time)

(this was the bug where if you had a tank PAW open and changed engine configuration to a different set of propellants, a new _extra_ button would be added to the tank PAW for that engine instead of replacing the existing button. If you changed configurations on that engine 20 times then the tank PAW would get 20 extra buttons until you closed the PAW and opened it again.