Expose propellantStability directly without rounding
Introduces GetPropellantProbability/GetUllageProbability to get
the rounded version that forces veryStable to 100% (couldn't think
of a better name, even though this isn't really the proper probability
since the exponent isn't applied).

This lets MJ 'see' values between 0.996 and 1.0

Didn't get this compiling yet because of all the RF deps and ROUtils in
particular so be a little careful before merging.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>